+++ demisto_sdk/commands/validate/validate_manager.py
@@ -53,6 +53,7 @@ from demisto_sdk.commands.common.hook_validations.integration import \
     IntegrationValidator
 from demisto_sdk.commands.common.hook_validations.layout import (
     LayoutsContainerValidator, LayoutValidator)
+from demisto_sdk.commands.common.hook_validations.lists import ListsValidator
 from demisto_sdk.commands.common.hook_validations.mapper import MapperValidator
 from demisto_sdk.commands.common.hook_validations.pack_unique_files import \
     PackUniqueFilesValidator
@@ -533,6 +534,9 @@ class ValidateManager:
         elif file_type == FileType.PRE_PROCESS_RULES:
             return self.validate_pre_process_rule(structure_validator, pack_error_ignore_list)
 
+        elif file_type == FileType.LISTS:
+            return self.validate_lists(structure_validator, pack_error_ignore_list)
+
         elif file_type == FileType.DASHBOARD:
             return self.validate_dashboard(structure_validator, pack_error_ignore_list)
 
@@ -888,6 +892,12 @@ class ValidateManager:
                                                               json_file_path=self.json_file_path)
         return pre_process_rules_validator.is_valid_pre_process_rule(validate_rn=False, id_set_file=self.id_set_file,
                                                                      is_ci=self.is_circle)
+
+    def validate_lists(self, structure_validator, pack_error_ignore_list):
+        lists_validator = ListsValidator(structure_validator, ignored_errors=pack_error_ignore_list,
+                                         print_as_warnings=self.print_ignored_errors,
+                                         json_file_path=self.json_file_path)
+        return lists_validator.is_valid_list()
 
     def validate_dashboard(self, structure_validator, pack_error_ignore_list):
         dashboard_validator = DashboardValidator(structure_validator, ignored_errors=pack_error_ignore_list,