+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
-# 1.5.2
+
+* added a new command **coverage-analyze** to generate and print coverage reports.
 * Fixed an issue in **validate** in repositories which are not in GitHub or GitLab
 * Added a validation that verifies that readme image absolute links do not contain the working branch name.
+* Added support for List content item in the **format**, **validate**, **download**, **upload**, **create-id-set**, **find-dependecies** and **create-content-artifacts** commands.
 * Added a validation to ensure reputation command's default argument is set as an array input.
+* Added the `--fail-duplicates` flag for the **merge-id-set** command which will fail the command if duplicates are found.
+* Added the `--fail-duplicates` flag for the **create-id-set** command which will fail the command if duplicates are found.
 * Added a validation to avoid contributors using fromlicense key in configuration parameters in integration yml.
 
 # 1.5.1